adds mapping of http status to no report, cloned report, frs started report adds transforms from FrsApi to Entity classes adds frs api configuration adds swagger plug-in to generated FrsApi and model classes modifies ui to recognize http status to create emissions report (step 1) modifies ui to create facility site (step 2)

+++ cef-web/src/main/java/gov/epa/cef/web/service/EmissionsReportService.java
@@ -1,11 +1,11 @@
 package gov.epa.cef.web.service;
 
-import java.util.List;
-
 import gov.epa.cef.web.domain.EmissionsReport;
 import gov.epa.cef.web.exception.ApplicationException;
 import gov.epa.cef.web.service.dto.EmissionsReportDto;
 import net.exchangenetwork.wsdl.register.program_facility._1.ProgramFacility;
+
+import java.util.List;
 
 public interface EmissionsReportService {
 
@@ -15,7 +15,7 @@ public interface EmissionsReportService {
      * @return
      */
     List<EmissionsReportDto> findByFacilityEisProgramId(String facilityEisProgramId);
-    
+
 
 
     /**
@@ -34,24 +34,24 @@ public interface EmissionsReportService {
 
     /**
      * Find the most recent report for a given facility
-     * @param facilityEisProgramId {@link ProgramFacility}'s programId 
+     * @param facilityEisProgramId {@link ProgramFacility}'s programId
      * @return
      */
     EmissionsReportDto findMostRecentByFacilityEisProgramId(String facilityEisProgramId);
 
-    
+
     String submitToCromerr(Long emissionsReportId, String activityId) throws ApplicationException;
 
 
     /**
      * Create a copy of the emissions report for the current year based on the specified facility and year.  The copy of the report is NOT saved to the database.
-     * @param facilitySiteId 
+     * @param facilityEisProgramId
      * @param currentReportYear The year of the report that is being created
      * @return
      */
-    EmissionsReport createEmissionReportCopy(String facilityEisProgramId, Integer currentReportYear);
+    EmissionsReport createEmissionReportCopy(String facilityEisProgramId, short currentReportYear);
 
-    
+
     /**
      * Save the emissions report to the database.
      * @param emissionsReport
